+++ core/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./codex.js";
 export * from "./files.js";
 export * from "./routing.js";
 export * from "./serial.js";